+++ config/utils.go
@@ -1,6 +1,10 @@
 package config
 
-import "strconv"
+import (
+	"fmt"
+	"strconv"
+	"strings"
+)
 
 // Check if string array contains string
 func contains(slice []string, sa string) bool {
@@ -19,6 +23,12 @@ func getMostlyPrioriesConfigKey(option map[string]interface{}) interface{} {
 			return *(value.(*int))
 		} else if option["type"] == "bool" {
 			return *(value.(*bool))
+		} else if option["type"] == "string_array" {
+			tmp := []string{}
+			for _, item := range *(value.(*StringArrayFlag)) {
+				tmp = append(tmp, item)
+			}
+			return tmp
 		}
 		return *(value.(*string))
 	} else if value, ok := option["file"]; ok {
@@ -27,6 +37,13 @@ func getMostlyPrioriesConfigKey(option map[string]interface{}) interface{} {
 			return i
 		} else if option["type"] == "bool" {
 			return value.(string) == "true"
+		} else if option["type"] == "string_array" {
+			valueSlice := value.([]interface{})
+			stringSlice := []string{}
+			for _, v := range valueSlice {
+				stringSlice = append(stringSlice, v.(string))
+			}
+			return stringSlice
 		}
 		return value
 	} else if value, ok := option["env"]; ok {
@@ -35,8 +52,34 @@ func getMostlyPrioriesConfigKey(option map[string]interface{}) interface{} {
 			return i
 		} else if option["type"] == "bool" {
 			return value.(string) == "true"
+		} else if option["type"] == "string_array" {
+			return strings.Split(value.(string), ",")
 		}
 		return value
 	}
+
+	// Send back default
+	if option["type"] == "string_array" {
+		fmt.Println(option["default"].(*StringArrayFlag))
+		tmp := []string{}
+		for _, item := range *(option["default"].(*StringArrayFlag)) {
+			tmp = append(tmp, item)
+		}
+		return tmp
+	}
 	return option["default"]
+}
+
+// StringArrayFlag is a string array for flag
+type StringArrayFlag []string
+
+// String return array as string
+func (i *StringArrayFlag) String() string {
+	return ""
+}
+
+// Set a item to array
+func (i *StringArrayFlag) Set(value string) error {
+	*i = append(*i, value)
+	return nil
 }